b2client: Add the ability to use a restricted Application ID
B2 allows you to create Application IDs that are restricted to specific
buckets. To allow for this, read back the information provided by the
authorization request and use the allowed bucket permissions if they are
found.

An account-level Application ID will still continue to function
normally, as would an Application ID that has access to all buckets.