Add dependency on github.com/cenkalti/backoff
Used by RetryableStorage implementation to provide
exponential backoff for storage operation retries.

A version change in `dep` caused the formatting of
`Gopkg.lock` to change significantly. I manually
reformatted so that only the non-whitespace changes
are part of this commit (to keep the PR shorter).

Before merging, I'd argue that `dep ensure` should be
run with the latest or project-adopted version of `dep`
so that my manual formatting is reverted to `dep`
default behavior.